+++ app/src/main/java/com/goldadorn/main/activities/cart/ICartData.java
@@ -9,9 +9,9 @@ import java.util.List;
  * Created by Kiran BH on 27/03/16.
  */
 public interface ICartData {
-    public void storeCartData(ArrayList<Product> cart, double costTotal);
+    public void storeCartData(ArrayList<Product> cart, float costTotal);
     public List<Product> getCartProducts();
-    public double getBillableAmount();
+    public float getBillableAmount();
     public void setPaymentDone(boolean done);
     public boolean isPaymentDone();
 }
